# Optimize Naive K-means with OpenMP