Following up on the Hamerly K-means optimization, I've also applied OpenMP to the naive K-means implementation. Here are the results.

**Benchmark Results**
I ran a benchmark with the following settings:
```cpp
Number of data points: 10,000
Number of dimensions: 50
Number of clusters: 10
Maximum iterations: 10
```
**Here's what I found:**
```
Original version: 13.753 seconds
OpenMP version: 12.8868 seconds
```

This represents a **6.3% decrease** in execution time.
Looking forward to your thoughts :)