@conradsnicta 2024-07-11
The speedup is definitely off.  See https://doi.org/10.1109/ICSPCS.2017.8270510 for an alternative implementation (which is part of Armadillo: https://arma.sourceforge.net/docs.html#kmeans).