@rcurtin 2024-07-12
Have you checked whether the distance computation itself is parallelized?  If it's not, simply parallelizing that instead might give significantly more speedup.  (Note that mlpack doesn't require OpenMP to be available or enabled, so going beyond `#pragma omp parallel for` can be a little bit tricky, if you want to go the route of the paper @conradsnicta linked to, but it can definitely be done.)