@conradsnicta 2024-07-15
@MarkFischinger I don't think hardcoding the block size is the way to go.  It would be better to divide the available data into segments based on the number of available threads, which is determined at run-time.

The nominal segment size can be simply `number_of_data_vectors / number_of_threads` in integer form. The last segment will need to have a possibly different size than the nominal size, to take into account that the division may not be clean.

As an example, say we have 123 vectors and 4 threads, resulting in a nominal segment size of 30.  The first 3 segments would each have a size of 30, thereby taking up 90 vectors in total. The last segment size would then be 123 - 90 = 33. 

A bunch of corner cases would also need to be taken care of, such as ensuring a minimum number of vectors per thread, and having more threads than vectors.