@MarkFischinger 2024-07-22
I fixed the error and implemented a new `minVectorsPerThread` variable. Here are the new benchmarks:
```console
mark@mark:~/gsoc/mlpack-benchmarks/src/kmeans$ ./naive_kmeans
Running benchmark for naive KMeans...
Benchmark for naive KMeans completed.
Total time: 7.37996 seconds
```