@MarkFischinger 2024-08-20
@rcurtin I updated the code following the feedback and optimized the distance calculation and OpenMP parallelization. Below are the additional benchmarks comparing the performance before and after these changes:

**Before Optimization:**

```
Running benchmark for naive KMeans...
Benchmark for naive KMeans completed.
Total time: 11.494 seconds
Running test for KMeans...
Assignments:         1        1        0        0

Centroids:
   5.5000   1.5000
   5.5000   1.5000

Test passed successfully!
```

**After Optimization:**
```
Running benchmark for naive KMeans...
Benchmark for naive KMeans completed.
Total time: 6.35895 seconds
Running test for KMeans...
Assignments:         1        1        0        0

Centroids:
   5.5000   1.5000
   5.5000   1.5000

Test passed successfully!
```

I went back to the optimized Euclidean distance formula, precomputing squared norms for both centroids and data points. This cuts down on redundant calculations, particularly when dealing with high-dimensional data. I also improved the OpenMP parallelization by making the parallel loops more fine-grained. Using a static schedule helps minimize synchronization overhead, leading to better thread utilization and faster execution. Finally, I focused on memory and cache efficiency by reusing precomputed values and that the data access is more cache friendly.